fix an issue in node environment that btoa is undefined, by keeping SightCard component in root file
